requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
cryptography==45.0.3
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.2.1
django-browser-reload==1.18.0
django-ckeditor-5==0.2.18
django-htmx==1.23.0
django-import-export==4.3.7
django-js-asset==3.1.2
django-phonenumber-field==8.1.0
django-tailwind==4.0.1
django-unfold==0.63.0
Faker==37.3.0
idna==3.10
oauthlib==3.2.2
phonenumbers==9.0.6
pillow==11.2.1
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-social-auth==0.3.6
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
